Add native kernel extensions with ASM helpers and fix sim output
Add 4 new native handlers (arraycopy, memcmp, writeBytes, stringFromBytes)
to picoJVM core with hand-written 8085 ASM implementations for the sim
target. ASM handlers are 5x smaller than compiler output (257B vs 1,286B)
by using direct memory access and tight register loops. Net .text increase
is +1,784B over baseline (25,705B total).

Also fix sim output bug: output_ptr at 0x0200 collided with BSS, and ASM
helper struct offsets were stale after PJVMFrame grew to 11 bytes (pc
uint16_t→uint32_t). Updated OFF_PC/OFF_CUR_CB/OFF_SP/OFF_CUR_LB.

Platform package (Console, Arrays, IO) and test coverage included.

+++ platform/generic.c
@@ -65,11 +65,11 @@ void pjvm_platform_putchar(uint8_t ch) {
     if (pjvm_putchar_fn) pjvm_putchar_fn(ch);
 }
 
-uint8_t pjvm_platform_peek8(uint16_t a) {
-    return heap_mem[a];
+uint8_t pjvm_platform_peek8(uint32_t a) {
+    return heap_mem[(uint16_t)a];
 }
 
-void pjvm_platform_poke8(uint16_t a, uint8_t v) {
+void pjvm_platform_poke8(uint32_t a, uint8_t v) {
     heap_mem[a] = v;
 }
