+++ platform/i8085_target.c
@@ -48,11 +48,11 @@ void pjvm_platform_putchar(uint8_t ch) {
     OUTPUT_PORT = ch;
 }
 
-uint8_t pjvm_platform_peek8(uint16_t a) {
-    return *(uint8_t *)(uintptr_t)a;
+uint8_t pjvm_platform_peek8(uint32_t a) {
+    return *(uint8_t *)(uintptr_t)(uint16_t)a;
 }
 
-void pjvm_platform_poke8(uint16_t a, uint8_t v) {
+void pjvm_platform_poke8(uint32_t a, uint8_t v) {
     *(uint8_t *)(uintptr_t)a = v;
 }
